add admin and multiple users validation

+++ src/main/java/org/lessons/java/photoalbum/controller/PhotoController.java
@@ -5,9 +5,11 @@ import java.util.Optional;
 
 import org.lessons.java.photoalbum.db.pojo.Category;
 import org.lessons.java.photoalbum.db.pojo.Photo;
+import org.lessons.java.photoalbum.db.pojo.User;
 import org.lessons.java.photoalbum.db.serv.CategoryService;
 import org.lessons.java.photoalbum.db.serv.PhotoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -31,26 +33,43 @@ public class PhotoController {
 	
 	@GetMapping
 	public String photos__index(Model model,
-			@RequestParam(required = false) String title) {
-		List<Photo> photos = title == null 
-				? photoService.findAll()
-				: photoService.findByTitle(title);
+			@RequestParam(required = false) String title,
+			Authentication authentication) {
+		User user = (User) authentication.getPrincipal();
 		
-		model.addAttribute("photos", photos);
-		
+		if(user.getId() == 1) {
+			List<Photo> photos = title == null 
+					? photoService.findAll()
+					: photoService.findByTitle(title);
+			
+			model.addAttribute("photos", photos);
+		}else {
+			List<Photo> photos = title == null 
+					? photoService.findByUser(user)
+							: photoService.findByUserAndTitle(user, title);
+			
+			model.addAttribute("photos", photos);
+		}
 		return "photos/photos_index";
 	}
 	
 	@GetMapping("/{photo_id}")
 	public String photos__show(Model model, 
-			@PathVariable int id) {
+			@PathVariable int id,
+			Authentication authentication) {
+		User user = (User) authentication.getPrincipal();
+		
 		Optional<Photo> photoOpt = photoService.findById(id);
 		if(!photoOpt.isEmpty()) {
+			
 			Photo photo = photoOpt.get();
 			
-			model.addAttribute("photo", photo);
+			if(user.getId() == photo.getUser().getId()) {				
+				model.addAttribute("photo", photo);
+				
+				return "photos/photo_show";
+			}
 			
-			return "photos/photo_show";
 		}
 		return "not_found";
 	}
@@ -69,13 +88,16 @@ public class PhotoController {
 	@PostMapping("/create")
 	public String photos__store(Model model,
 			@Valid @ModelAttribute Photo photo,
-			BindingResult bindingResult) {
+			BindingResult bindingResult,
+			Authentication authentication) {
+		User user = (User) authentication.getPrincipal();
 		if(bindingResult.hasErrors()) {
 			System.err.println("Error: ");
 			bindingResult.getAllErrors().forEach(System.err::println);
 			return "photos/photos_form";
 		}else {
 			try {
+				photo.setUser(user);
 				photoService.save(photo);
 			}catch(Exception e) {
 				System.err.println(e.getMessage());
@@ -87,16 +109,35 @@ public class PhotoController {
 	
 	@GetMapping("/update/{photo_id}")
 	public String photos__edit(Model model,
-			@PathVariable("photo_id") int id) {
+			@PathVariable("photo_id") int id,
+			Authentication authentication) {
+		User user = (User) authentication.getPrincipal();
+		
+		if(user.getId() == 1) {
+			Optional<Photo> photoOpt = photoService.findById(id);
+			if(!photoOpt.isEmpty()) {
+				Photo photo = photoOpt.get();
+				List<Category> categoriesList = categoryService.findAll();
+				
+				model.addAttribute("categoriesList", categoriesList);
+				model.addAttribute("photo", photo);
+				
+				return "photos/photos_form";
+			}
+		}
+		
 		Optional<Photo> photoOpt = photoService.findById(id);
 		if(!photoOpt.isEmpty()) {
 			Photo photo = photoOpt.get();
-			List<Category> categoriesList = categoryService.findAll();
 			
-			model.addAttribute("categoriesList", categoriesList);
-			model.addAttribute("photo", photo);
-			
-			return "photos/photos_form";
+			if(user.getId() == photo.getUser().getId()) {					
+				List<Category> categoriesList = categoryService.findAll();
+				
+				model.addAttribute("categoriesList", categoriesList);
+				model.addAttribute("photo", photo);
+				
+				return "photos/photos_form";
+			}
 		}
 		
 		return "not_found";
@@ -106,15 +147,25 @@ public class PhotoController {
 	public String photos__update(@Valid @ModelAttribute Photo photo,
 			BindingResult bindingResult,
 			Model model,
-			@PathVariable("photo_id") int id) {
+			@PathVariable("photo_id") int id,
+			Authentication authentication) {
 		if(bindingResult.hasErrors()) {
 			System.err.println("Error: ");
 			bindingResult.getAllErrors().forEach(System.err::println);
 			return "photos/photos_form";
 		}else {
 			try {
-				photo.setId(id);
-				photoService.save(photo);
+				Photo oldphoto = photoService.findById(id).get();
+				User user = oldphoto.getUser();
+				User authUser = (User) authentication.getPrincipal();
+				
+				if(authUser.getId() == 1) {					
+					photo.setUser(user);
+				}else {
+					photo.setUser(authUser);
+				}
+				
+				photo.setId(id);				photoService.save(photo);
 			}catch(Exception e) {
 				System.err.println(e.getMessage());
 				return "photos/photos_form";
@@ -125,12 +176,16 @@ public class PhotoController {
 	
 	@PostMapping("/delete/{photo_id}")
 	public String photos__delete(Model model, 
-			@PathVariable("photo_id") int id) {
+			@PathVariable("photo_id") int id,
+			Authentication authentication) {
+		User user = (User) authentication.getPrincipal();
 		Optional<Photo> photoOpt = photoService.findById(id);
 		if(!photoOpt.isEmpty()) {
 			Photo photo = photoOpt.get();
-			photoService.deletePhoto(photo);
-			return "redirect:/photos";
+			if(user.getId() == photo.getUser().getId()) {					
+				photoService.deletePhoto(photo);
+				return "redirect:/photos";
+			}
 		}
 		return "not_found";
 	}
